miniLM/requirements.txt
# MiniChat (MiniLM) Dependencies

# Web UI
streamlit>=1.28.0

# LLM Integration
ollama>=0.1.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# LangChain (for document processing utilities)
langchain>=0.1.0
langchain-community>=0.0.10

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11

# Desktop App
pywebview>=4.0.0

# Testing
pytest>=7.0.0
hypothesis>=6.0.0
